requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.0.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
async-timeout=4.0.2=pypi_0
attrs=21.4.0=pypi_0
blas=1.0=mkl
bottleneck=1.3.4=py310h9102076_0
brotli=1.0.9=h166bdaf_7
brotli-bin=1.0.9=h166bdaf_7
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.6.15=ha878542_0
cachetools=5.1.0=pypi_0
certifi=2022.6.15=py310hff52083_0
cffi=1.15.0=py310hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=37.0.1=py310h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd8ed1ab_0
dbus=1.13.18=hb2f20db_0
expat=2.4.8=h27087fc_0
ffmpeg=4.3=hf484d3e_0
filelock=3.7.0=pypi_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.33.3=py310h5764c6d_0
freetype=2.11.0=h70c0345_0
frozenlist=1.3.0=pypi_0
fsspec=2022.5.0=pypi_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.6.6=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.46.3=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
huggingface-hub=0.6.0=pypi_0
icu=58.2=hf484d3e_1000
idna=3.3=pyhd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.1.0=pypi_0
jpeg=9e=h7f8727e_0
kiwisolver=1.4.2=py310h295c915_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_0
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=11.2.0=h1234567_0
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_0
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h2818925_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.15=h7f8727e_0
libxml2=2.9.14=h74e7548_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.7=pypi_0
matplotlib=3.5.1=py310h06a4308_1
matplotlib-base=3.5.1=py310ha18d171_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
multidict=6.0.2=pypi_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
numexpr=2.7.3=py310hd732450_1
numpy=1.22.3=py310hfa59a62_0
numpy-base=1.22.3=py310h9585f30_0
oauthlib=3.2.0=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1o=h7f8727e_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.2=py310h295c915_0
pcre=8.45=h9c3ff4c_0
pillow=9.0.1=py310h22f2fdc_0
pip=21.2.4=py310h06a4308_0
protobuf=3.20.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydeprecate=0.3.2=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.9.2=py310h295c915_6
pysocks=1.7.1=py310h06a4308_0
python=3.10.4=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.10=2_cp310
pytorch=1.11.0=py3.10_cuda11.3_cudnn8.2.0_0
pytorch-lightning=1.6.3=pypi_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=pypi_0
qt=5.9.7=h5867ecd_1
readline=8.1.2=h7f8727e_1
regex=2022.4.24=pypi_0
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scikit-learn=1.1.1=pypi_0
scipy=1.7.3=py310hfa59a62_0
setuptools=61.2.0=py310h06a4308_0
sip=4.19.13=py310h295c915_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.3=hc218d9a_0
tensorboard=2.9.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.11=h1ccaba5_1
tokenizers=0.12.1=pypi_0
torchaudio=0.11.0=py310_cu113
torchmetrics=0.8.2=pypi_0
torchvision=0.12.0=py310_cu113
tornado=6.1=py310h5764c6d_3
tqdm=4.64.0=pypi_0
transformers=4.19.2=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
unicodedata2=14.0.0=py310h5764c6d_1
urllib3=1.26.9=py310h06a4308_0
werkzeug=2.1.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
yarl=1.7.2=pypi_0
zlib=1.2.12=h7f8727e_2
zstd=1.5.2=ha4553b6_0
